refactor: add new components and update app structure
Introduce Button, Header, and Input components under src/components.
Refactor the project structure to move the main App logic to a Home
page, and update imports accordingly in main.tsx.
